requirements.txt
gradio
openai
ollama
python-dotenv
fastmcp
httpx
chromadb
langchain
langchain-text-splitters
markitdown[all]